+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 2.1.0
+
+- added CSS body tag that differentiates between R.Native (`.is-sg-codebase-v2`) & old Swift/Java based App (`.is-sg-codebase-v1`) 
+
 # 2.0.2
 
 - Fixed checkout order conversion data feed